BUG-187: make Byte and Short distinct primitives for overload resolution (#5665)
* BUG-184: make Byte and Short distinct primitives for overload resolution


* Fix CI regressions: parse Byte/Short library signatures and type primitive-array factory literals

- LibraryMetadataParsing.swift: parse 'B' and 'S' type-signature tokens for Byte/Short.
- HeaderHelpers+SyntheticArrayStubs.swift: make byteArrayOf/shortArrayOf element types Byte/Short.
- HeaderHelpers+SyntheticListConversionMembers.swift: toList() element types for ByteArray/ShortArray.
- CallTypeChecker+LambdaReturnTypeOverload.swift: include Byte/Short in uniformNumericLiteralParameterType and handle unary +/- over int literals so byteArrayOf/shortArrayOf can accept negative/fit constants.
- Add LibraryMetadataSignatureParsingTests regression test for B/S signature parsing.
- Add Scripts/diff_cases/byte_short_array.kt regression diff case.
- Update affected Sema golden files.


* Fix BUG-187: narrow int literals in member calls and indexed assign for Byte/Short

- Re-infer unsuffixed int literals and unary +/- over int literals against the
  surviving candidates' parameter type in prepareCallArguments, so member calls
  like shortArray.binarySearch(20) and byteArray.foo(1) narrow to Short/Byte.
- Filter set-value expected types to concrete wideable primitives in
  inferIndexedAssignExpr to avoid generic Map/ MutableMap set inference regressions
  while still narrowing ByteArray/ShortArray indexed assignments.
- Extend byte_short_array.kt with indexed assign + binarySearch regression coverage.
- Add ByteShortOverloadResolutionTests.testByteAndShortArrayLiteralNarrowing.
- Renumber completed TODO entry from duplicate BUG-186 to BUG-187.


* Update ComparisonSyntheticTopLevelTests for distinct Byte/Short maxOf/minOf resolution

With Byte/Short as distinct primitives, maxOf/minOf calls with Byte/Short
arguments resolve to the generic Comparable<T> overloads instead of the
Int fixed-arity special-call path. Update expectations accordingly and drop
unsubstituted signature assertions that are invalid for generic overloads.


* Update AnnotationSemanticTests for distinct Byte primitive in function types

With Byte as a distinct primitive,  and
 now correctly report  parameters
instead of . Update expectations in
.


* Fix CI regressions: elvis literal narrowing, Byte/Short until matrix, and test expectations


* Add BUG-207 TODO entry for Byte/Short distinct primitives


---------

+++ Tests/CompilerCoreTests/Sema/StringToShortOrNullFunctionTests.swift
@@ -10,8 +10,9 @@ import Testing
 ///   runtime symbol `kk_string_toShortOrNull` declared in
 ///   `Sources/RuntimeABI/RuntimeABISpec+String.swift`.
 /// - The extension resolves cleanly from source code and produces no Sema
-///   diagnostics for a call returning `Int?` (Short is widened to Int in ABI).
-/// - An elvis fallback on the nullable result type-checks correctly.
+///   diagnostics for a call returning `Short?`.
+/// - An elvis fallback on the nullable result type-checks correctly, narrowing
+///   the fallback integer literal to `Short`.
 @Suite
 struct StringToShortOrNullFunctionTests {
     @Test
@@ -21,7 +22,7 @@ struct StringToShortOrNullFunctionTests {
             return raw.toShortOrNull()
         }
 
-        fun probe(): Int {
+        fun probe(): Short {
             val parsed = "32767".toShortOrNull()
             return parsed ?: 0
         }
